Add files via upload
The loan providing companies find it hard to give loans to the people due to their insufficient or non-existent credit history. Because of that, some consumers use it as their advantage by becoming a defaulter. Suppose we work for a consumer finance company which specialises in lending various types of loans to urban customers. I used EDA to analyse the patterns present in the data. This will ensure that the applicants capable of repaying the loan are not rejected.